rename src to dune_client (only way I can get it to work)
