+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = dune-client
+name = dune_client
 description = A simple framework for interacting with Dune Analytics official API service.
 long_description = file: README.md
 long_description_content_type = text/markdown; charset=UTF-8
@@ -18,11 +18,6 @@ project_urls =
 [options]
 zip_safe = False
 packages=find:
-exclude =
-  tests
-  tests.*
-  example
-  example.*
 platforms = any
 install_requires =
   duneapi==4.0.0
@@ -33,5 +28,13 @@ install_requires =
 python_requires = >=3.9
 setup_requires =
     setuptools_scm
-# package_data=
-#  dune-client=py.typed
+
+[options.packages.find]
+exclude =
+  tests
+  tests.*
+  example
+  example.*
+
+[options.package_data]
+dune_client = py.typed
